Updated configs for powerline; forced 256 colors in xterm and tmux
Although I'm quite fond of my custom statusline, the idea of a common
interface and API to extend and theme the statusline is appealing. Also
powerline looks nicer than mine. So I'm giving it a shot.

Generally I dislike having to force 256 colors and would much prefer if
xterm and tmux could intelligently detect and use (similar to how 8 and
16 colors are handled). Powerline is toying with adding support for 16
colors, hopefully they'll get it working. Perhaps I need to add a
detect-and-workaround for powerline if Vim doesn't start with 256
colors.