Move git editc to standalone script and populate quickfix
Why: Editing files from a commit is nice but it's easy to forget where
the edits were made. This is a simple way to populate the quickfix list
with the line numbers of the start of each hunk. Sort of a middle ground
between a diff view and a regular file edit.