Generalize fetchall; add ssh_fetchall for working with GitHub/GitLab
This doesn't do anything intelligent so you'll want to run it in
a directory of repos cloned from just one provider or the other to take
advantage of the individual pipelining.

Probably wouldn't be too hard to fetch remote information from each repo
upfront, dedupe, and establish connections to each one before kicking
off the bulk fetches. Might be cool. But this is already a silly
function in the first place and that will make it that much more
silly...