Factor out find from grep in gext wrapper for standalone find
Why: It's useful to be able to use vanilla find and still exclude common
dirs. This first began as a way to reuse exclude patterns in both find
and grep but grep removed GREP_OPTIONS and I moved all the filtering to
just the find. We no longer needed to keep GREP_OPTIONS up-to-date when
cd'ing around the file system. This change finalizes that separation and
allows usage as just a standalone find, thus the find-specific prune
renames.