Add stripped down vim-tmux-navigator implementation
I try to avoid these kinds of hacks but this is a pretty minimal one.
This allows me to use tmux keys to jump around between tmux splits as
well as Vim windows without having to make the distinction which is
really nice when you're on a large screen with lots of panes and splits.
Ideally this would be done all within a single terminal multiplexer and
not in an editor (ala Vis or Kakoune), but this will do for now.