Remove 'cursor' from zsh-syntax-highlighting default highlighters
This setting has started changing the cursor from my terminal default
when I switch vi-mode modes which is not something I want. I use block
cursor because it's easy to spot; underline isn't. That behavior is not
documented -- in fact the docs don't explain what this option does at
all other than 'the style for the current cursor position' so I'm not
sure why I added it in the first place.