Improve performance when typing query
This change allows the prompt to keep up with the user typing quickly.

The search performance starts to be slow at the tens of thousands and is
too slow to use at the hundreds of thousands. Perhaps this can be
improved by optmizing the lines/columns calculations. Needs benchmarks.

FIXME: this seems to be truncating very long lists.