Add a simple file/xargs based job queue
I often want to download a list of things over slow connections so batches of a
few at a time are best to use available bandwidth but without saturation and
then move on to the next item(s) when done. There's a lot of tools to do this
but I'd prefer to just run a shell watcher than a background daemon or
something more complex/featurefull.