Add AddKeysToAgent and remove all identities on screensaver activation
This setting makes it so when you type `ssh` and it asks you for the
passphrase for a key the identity is also added to the running
ssh-agent. No more having to ctrl-c, add the identity, then ssh again.
This one-step process to both unlock an identity and ssh somewhere makes
it seamless enough that I can finally remove unlocked identities
whenever my screensaver activates. Keeping them unlocked in memory for
long-term periods is an unnecessary risk.