Remove Nethack prompt
I'm a couple months into the experiment to move status info from my
prompt to a tmux pane status display and still enjoying it. The Nethack
prompt was a lot of fun to put together but I've gradually moved to more
and more minimal prompts so I think it's time to archive this to
memories and Git history.

...
.d.
...

See you in the Mazes of Menace, buddy.