Remove Vim + tmux navigation integration
In theory it was nice to not care if you're in a tmux pane or a Vim
window and simply navigate with the same keys regardless. In practice
sometimes I run Vim without vimrc/plugins which breaks this, and it's
annoying when there's not an exact 1-to-1 when say a movement command
wraps around the pane and choosing the top/bottom or left/right window
is ambiguous. After experimentation I'd prefer to stick with Vim
movement for Vim windows and tmux movement for tmux panes.