Pass the line numbers in the selected range to the wrapped function
Also update join and surround plugins to take the additional param.

This isn't my favorite approach. Overloading the opfuncwrapper to
accomodate both edits and non-edits is an anti-pattern. But given that
opfunc needs to be a string I'm currently stumped on a better way. Maybe
someone else, or future me, will see a better path.