Add multi-buffer :update command and change inconsistent :wa
I'm trying to get into the habit of using :up instead of :w to only
write the file when there are new changes to write. (Useful for, say,
running autocommands after the file is written to, say, compile the
file.

Changing builtin Vim commands is probably a bad idea but I'm going to
try this one on for size to see how it goes.