Remove -m in git show alias
Without -m you can see the resolution of a merge conflict. The only
changes in the output will be the conflicts that were resolved. With -m
you can see the full diff of a merge commit. I can't think of a good
alias for -m so I'll try to remember to pass it in manually when I want
it because usually I just want to see the conflict resolution.