Remove unused mappings that reuse the path of another file
I never remember these and thus never use them. It's often easier to cd
to a directory and then just run regular shell commands from there.
Let's try doubling-down on that workflow to see how well it works.

This also updates mappings that change directory to be window-specific
changes instead of global.

Finally, for mappings that I rarely use (and thus remember) but that do
come in handy on occasion it may be better to implement them as custom
commands instead of something with an obscure keyboard shortcut. That
gives a name to a thing which is hopefully more mnemonic and memorable.
I'll be experimenting with doing this more.