Switch several rarely used mappings to custom commands
I use these but rarely and I often need to look the mapping up first,
which kind of defeats the purpose of adding it to the vimrc in the first
place. Custom commands are a bit more discoverable and mnemonic so let's
try this out to see how it goes.