Replace <C-X><C-F> completion with fuzzy-finder completion
This seems to work ok so far. I don't love the dep on coreutils but I do
like the more muscle-memory friendly shell util. The Vim-only approach
worked well but was a little hacky and I want to start leaning on shell
utils for more things. Let's try this and see how it goes.