Revert "Replace <C-X><C-F> completion with fuzzy-finder completion"
This reverts commit dd5f185bc9cc7d10ffa7fdd92b0724ddcc1a8392.

This doesn't work well when starting to type a path and then triggering
completion from there. Perhaps that can be solved with <afile> and
somehow calculating a starting directory from <pwd> + <afile>.

And having to calculate a relative path from two absolute paths at the
end isn't ideal.

ftree may not be the right tool for this job. I'd prefer to skate around
the filesystem more quickly, without having to press enter on each
directory change.

Using an external tool may still be workable but I'd like to think on
this a bit more.