Have git frefs output as an un-executed git log command
Not sure why I didn't do this on the first pass. The idea is usually
I want to view a log of the git commit range, but sometimes it's a diff,
othertimes it's both, othertimes still it's something else entirely.
This makes it easy to do any of those; easier than a pastebuffer too.