Rewrite lorem to not use shuf internally
I don't recall why I used both heavy awk and also used shuf. Maybe
because solving this without shuf is hard and I got sick of working on
it? This change gives portability since now we're only depending on awk.
But the randomization is also less awesome because shuf is awesome and
because I'm using reservoire sampling to avoid a (maybe?) slower
multi-pass algorithm in awk, and because awk's randomization is not
awesome.