WIP: switch IPC communication from timeout to reset on user signal
Continued experiment. The IPC communication before would listen for
a short period of time for new messages before continuing the draw
process. There's a short, unnoticable delay as it checks for new
messages and the sleep, once started, won't be interrupted by new
messages. If instead we only look for new messages upon receiving a USR1
signal we can also interrupt the sleep. The downside of this approach is
the writer process will block if the reader process doesn't respond to
the signal.

I'm saving both approaches as separate Git commits to more easily
compare the two if I wish to go back to not using signals.