Revert "Improve secure temp directory creation"
This reverts commit 278e168cd842454a71315c8f75a4d4b1eab4c9d7.

The m4 approach is not as portable as I first thought. Even recent OS
X versions are running a version of m4 too old for this trick.

The Bash FAQ breaks down _why_ the random name approach is secure:

> You're relying on the atomicity of mkdir for your security, not the
> obscurity of your random name.
>
> — https://mywiki.wooledge.org/BashFAQ/062

So we should add an exit code check here as well.