Add mapping to copy quickfix entries for the current file as locations
As I'm skating between quickfix entries I find I like to explore all the
entries in a single file, up and down, before moving on to the next
file. This is one way to constrain movement to just the current file.
I don't love it, but it's simple. Ideally I'd have ]q [q equivalent
mappings that used the existing quickfix list but bounded by the current
file. I may attempt that in the future and replace this.