Remove printf positional parameters
It seems positional arguments to printf are not as portable as
I expected. I need to do more reading on this. AFAICT, the positional
arguments to printf are only part of the POSIX spec for the C function
[1] and not for the /usr/bin/printf binary [2]. I'm not sure how that
translates to printf() in the awk specification (let alone the
implementations).

The busybox version of awk doesn't support positional arguments, and it
even doesn't implement the width specifier which seems like an amazingly
basic oversight. The Bash printf builtin doesn't support positional
arguments but Zsh and Ksh do. Mawk doesn't support positional arguments.

I'm fairly shocked that all those implementations aren't simply wrappers
around the underlying C function that already supports all the things.
Super bizarre. There must be more to the story here.

[1] https://pubs.opengroup.org/onlinepubs/9699919799/functions/printf.html
[2] https://pubs.opengroup.org/onlinepubs/9699919799/utilities/printf.html

This double for-loop and fake multi-dimentional array approach is much
uglier and less elegant than what it's replacing, but I don't want to
repeat the heading or the value in the code. Recalculating the width is
unnecessary but cheap. I plan to just dep on gawk and be done with it,
but this change does get it working under busybox so I'd like to retain
this version under version-control.