Force line buffering for grep
Since xargs is parallelizing grep the output won't be in order, which is
fine, however occassionally multiple grep processes will try and write
at the same time which will interleave the output lines from both
processes resulting in garbled/broken output.

xargs can't force line buffering, and there's no POSIX utils to enable
that (such as stdbuf in coreutils), but GNU grep has a flag. Ostensibly
there's a performance hit, but I haven't observed anything noteworthy so
far.

This flag won't be portable, of course, so I may want to try and detect
it somehow long-term but I'm going to commit it for now as both notes,
and to more easily run experiments on my other platforms.