Add mappings for a debugger (currently Vimspector)
Not sure how long-lived this will be. Vimspector is huge, and using text
editor panes to build a TUI feels like the wrong approach. There are
obvious advantages to having tight integration between a debugger and
a text editor, such previewing the current file and setting/removing
breakpoints on certain lines, however I think tight integration can be
done without building one inside the other and would be more inline with
the "UNIX as the IDE" philosophy. But there's not a lot of options out
there so I'll run with this for now to see how much I care.