Add lr and xe repos to my default clone
I try to stick to POSIX compatibility for my shell scripts, however
there are a few insufficiencies that aren't easy/possible to work
around. The big two are find and xargs. I rely on non-POSIX flags pretty
heavily (-0, -P, -r, -print0, -printf) and that usually means installing
the GNU suite (Busybox did copy a few of those flags but not all). But
it's kind of a crapshoot writing feature-detection all over is way too
verbose for simple scripts.

Why POSIX? For portability? For simplicity? Because ideologically I like
the Suckless philosophy? Maybe I just like doing it because it's harder
:-P. If I choose a custom userspace then I break portability. But
I write my dotfiles for ml; anything intended for an audience should be
broken out into a dedicated repo. (Note to self: do that for a few of
the scripts in bin.)

There's the question if I choose to install an extra util on top of
POSIX, then why not just install the GNU suite and be done with it?
I think it really is just a philosophical desire to avoid bloated tools
where possible. The bloat doesn't really affect me in any practical way,
but there's something appealing about a 2k SLOC find/ls replacement with
better ergonomics, or a 600 SLOC replacement for xargs. If the
apocalypse comes and we have to rebuild society, government, industry,
and POSIX (^_^), we need to retain the principles and fundamentals. All
the feature-creep will happen later (whether we want it to or not).

Jokes aside, I'd like to experiment more with running a minimal
userspace. Something like busybox with a few hand-picked extras.