Add/remove various Zsh options
auto_cd is sometimes nice for typos but doesn't come up enough to
matter.

extended_glob I just don't make use of.

inc_append_history is redundant with share_history. Docs specify you
should use one or the other. If there's a side-effect I haven't seen it
in many years of using both but I'll keep an eye out with this change.

mark_dirs with this enabled `ls -F foo*` will produce duplicate trailing
slashes which looks annoying. I'm honestly not sure if I use this option
so disabling to see if I notice.

path_dirs probably seemed like a good idea when I first added it, but
it's better to be explicit instead.

pushd_ignore_dups because I'm interested to quickly jump back to
a previous location and don't need to retain an exact history of my
directory changes.