Move debugger keys to the middle of the F-keys row
Turns out I use VimspectorContinue much more often than I expected so
it's nice to have that mapped to a single key press. I can't use F12
because I have that dedicated in tmux so this moves my other F-key
bindings to the left and right of the debugger F-keys and moves a couple
functions that I don't use too often to commands instead of maps.