Add jumping to a tag externally to Vim (via quickfix)
Starting an experiment to externalize tags usage (listing, filtering,
jumping). The builtin tags list doesn't do fuzzy-finding and is a tad
slow & verbose to use when you have more than one page of tag matches.
It would be nice if there was a better mechanism to integrate external
tag functionality into Vim. It would also be nice if we had more than
just the two quickfix and location lists to work with.