Use .TO/.FROM instead of a table
These are built in to the LETTER DOCTYPE that ships with mom. I was
getting a 'numeric overflow' error when trying to use two tables
back-to-back on newer versions of groff. groff-1.22.3 worked ok, not
sure if the error was in that version or in the newer versions. I spent
an hour or two trying to track down the error, then gave up.