Add formatter separate from formatprg
Slow realization: I really like gq for reformatting prose (docstrings,
comments, Markdown, etc), so it makes sense to have the code reformatter
be separate from the prose reformatter. Also, I'm not sure equalprg is
a better fit -- both equalprg and formatprg are meant to be able to
operate on subsets of the code, whereas code formatters are meant to
always and only operate on the entire file.