Replace filter/undo with getline/setline
There are all sorts of edge cases with creating an undo entry for
a filter command that may succeed or fail. If we don't use undojoin then
undoing the formatting moves the cursor to the first line; if we do use
undojoin and it fails we get an error about undoing an undo that doesn't
exist.

Rather than making the change and trying to undo it if it fails, it is
better to not make the change at all if it fails...and all it costs us
is storing two copies of the file in memory (o_O).